# api/openapi/crm-v1.yaml
openapi: 3.1.0
info:
  title: Property CRM — Client API
  version: 1.2.0
  description: |
    Contract for the web application and the Flutter mobile app.

    **Contract rules (binding on both sides):**
    - Consumers MUST tolerate unknown response fields and unknown values in
      enums marked OPEN. Enums marked CLOSED only change with a version bump.
    - All unsafe requests SHOULD send `Idempotency-Key`; replayed offline
      writes MUST send it plus `X-Offline-Replay: 1`.
    - State-machine resources are server-authoritative: a stale transition
      returns `409` with code `state_conflict` and the current resource in
      `problem.current`; the client reconciles and retries deliberately.
    - Claims are ONLINE-ONLY: never queued offline (`422 claims_online_only`).
    - Server countdowns (`ttl_expires_at`, hold expiry) are authoritative;
      client timers are cosmetic.
    - Mobile clients send `X-App-Platform` and `X-App-Version` on every
      request; `426` means hard upgrade wall (body has localised copy).
  contact:
    name: CRM team
  license:
    name: Proprietary — internal project contract
    identifier: LicenseRef-Proprietary
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /listings/{listingId}/viewing-slots:
    get:
      operationId: listViewingSlots
      tags: [appointments]
      summary: Browse open bookable slots before placing a hold
      description: |
        Slots are generated from property access rules (occupancy, minimum
        notice for tenanted properties, blackout windows) in the property's
        timezone. Browse-then-hold is the intended flow; a hold on a listed
        slot can still fail with `409 slot_conflict` if another viewer takes
        it first — slots are availability, not reservations.
      parameters:
        - $ref: '#/components/parameters/ListingId'
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time }, description: Window capped server-side (default 14 days) }
      responses:
        '200':
          description: Open slots, ascending.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [starts_at, ends_at]
                      properties:
                        starts_at: { type: string, format: date-time }
                        ends_at: { type: string, format: date-time }
                        kind: { type: string, enum: [private, open_house] }
        '404': { $ref: '#/components/responses/NotFound' }
  /me/portfolio:
    get:
      operationId: listPortfolio
      tags: [portfolio]
      summary: My tracked properties with valuation estimates
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Portfolio entries.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/PortfolioEntry' }
                  next_cursor: { type: [string, 'null'] }
    post:
      operationId: addPortfolioEntry
      tags: [portfolio]
      summary: Track a property (investor-entered figures)
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortfolioEntryInput' }
      responses:
        '201':
          description: Entry created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioEntry' }
        '409':
          description: Property already tracked (`portfolio_duplicate`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /me/portfolio/{propertyId}:
    patch:
      operationId: updatePortfolioEntry
      tags: [portfolio]
      summary: Update figures or status for a tracked property
      parameters:
        - { name: propertyId, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortfolioEntryUpdate' }
      responses:
        '200':
          description: Updated entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioEntry' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: removePortfolioEntry
      tags: [portfolio]
      summary: Stop tracking a property
      parameters:
        - { name: propertyId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Removed. }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /appointments/holds:
    post:
      operationId: createHold
      tags: [appointments]
      summary: Place a slot hold (TTL) before booking
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listing_id, starts_at, ends_at]
              properties:
                listing_id: { type: string, format: uuid }
                starts_at: { type: string, format: date-time }
                ends_at: { type: string, format: date-time }
      responses:
        '201':
          description: Hold placed; convert before `expires_at`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SlotHold' }
        '409':
          description: Slot conflict (`slot_conflict`) or notice-period violation (`min_notice`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /appointments:
    post:
      operationId: bookAppointment
      tags: [appointments]
      summary: Convert a hold into a booking (starts agent dispatch)
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hold_id]
              properties:
                hold_id: { type: string, format: uuid }
                notes: { type: string }
      responses:
        '201':
          description: Appointment created in `dispatching`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Appointment' }
        '410':
          description: Hold expired (`hold_expired`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: My appointments (any role)
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: state, in: query, schema: { type: string }, description: Comma-separated states }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of appointments.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Appointment' }
                  next_cursor: { type: [string, 'null'] }
  /appointments/{appointmentId}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Appointment detail
      parameters: [{ $ref: '#/components/parameters/AppointmentId' }]
      responses:
        '200':
          description: Appointment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Appointment' }
          headers:
            ETag: { schema: { type: string } }
        '404': { $ref: '#/components/responses/NotFound' }
  /appointments/{appointmentId}/cancel:
    post:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel (notice policy applied server-side)
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Cancelled; body includes the policy consequence applied.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CancellationResult' }
        '409': { $ref: '#/components/responses/StateConflict' }
...
  /appointments/{appointmentId}/register:
    post:
      operationId: registerOpenHouse
      tags: [appointments]
      summary: Register for an open house (capacity-aware)
      description: |
        Positions within the appointment's capacity are confirmed attendees;
        later registrants are waitlisted and promoted automatically (with a
        notification) when someone unregisters.
      parameters: [{ $ref: '#/components/parameters/AppointmentId' }]
      responses:
        '201':
          description: Registration result.
          content:
            application/json:
              schema:
                type: object
                required: [position, confirmed]
                properties:
                  position: { type: integer }
                  confirmed: { type: boolean, description: false = waitlisted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Not an open house (`not_open_house`) or wrong state.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /appointments/{appointmentId}/unregister:
    post:
      operationId: unregisterOpenHouse
      tags: [appointments]
      summary: Withdraw an open-house registration
      parameters: [{ $ref: '#/components/parameters/AppointmentId' }]
      responses:
        '204': { description: Removed; the next waitlisted viewer is promoted. }
        '404': { $ref: '#/components/responses/NotFound' }
  /appointments/{appointmentId}/check-in:
    post:
      operationId: checkIn
      tags: [appointments]
      summary: Attendance proof — check in (agent or viewer)
      parameters: [{ $ref: '#/components/parameters/AppointmentId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttendanceProofInput' }
      responses:
        '204': { description: Recorded. }
        '409': { $ref: '#/components/responses/StateConflict' }
        '422':
          description: Evidence rejected (`geofence_out_of_range`, `bad_code`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /appointments/{appointmentId}/check-out:
    post:
      operationId: checkOut
      tags: [appointments]
      summary: Attendance proof — check out
      parameters: [{ $ref: '#/components/parameters/AppointmentId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttendanceProofInput' }
      responses:
        '204': { description: Recorded. }
        '409': { $ref: '#/components/responses/StateConflict' }
  /appointments/{appointmentId}/feedback:
    post:
      operationId: appointmentFeedback
      tags: [appointments]
      summary: Structured post-visit feedback (agent or viewer)
      parameters: [{ $ref: '#/components/parameters/AppointmentId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppointmentFeedbackInput' }
      responses:
        '204': { description: Recorded. }
  /appointments/{appointmentId}/outcome:
    post:
      operationId: appointmentOutcome
      tags: [appointments]
      summary: Record the viewing outcome (agent)
      parameters: [{ $ref: '#/components/parameters/AppointmentId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome]
              properties:
                outcome: { type: string, enum: [interested, offer_intent, rejected, no_show_viewer] }
                notes: { type: string }
      responses:
        '204': { description: Recorded; routed into the demand pipeline. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued JWT (Authorization Code + PKCE for humans).
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
      description: Same key + same payload replays the stored response; same key + different payload → 409 `idempotency_key_reuse`. Expires after 24 h.
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    ListingId:
      name: listingId
      in: path
      required: true
      schema: { type: string, format: uuid }
    AppointmentId:
      name: appointmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: Resource not found or not visible to this role.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    StateConflict:
      description: Stale state transition; `problem.current` carries the current resource.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457. `code` is the stable machine contract; `current` appears on `state_conflict`.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        current: { type: object, description: Current resource on state_conflict }
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: string, description: 'Decimal string, e.g. "285000.00"' }
        currency: { type: string, minLength: 3, maxLength: 3 }
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number }
        lng: { type: number }
...
    PortfolioEntryInput:
      type: object
      required: [property_id, purchase_price, monthly_rental_income, monthly_expenses]
      properties:
        property_id: { type: string, format: uuid }
        purchase_price: { $ref: '#/components/schemas/Money' }
        monthly_rental_income: { $ref: '#/components/schemas/Money' }
        monthly_expenses: { $ref: '#/components/schemas/Money' }
        outstanding_debt:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: 'Remaining mortgage principal (a stock/balance). Optional; absent = paid off or cash purchase. Investor-entered, never estimated.'
        monthly_mortgage_payment:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: 'Monthly installment (a flow/payment). Optional; absent = paid off or cash purchase. Investor-entered, never estimated.'
        status: { type: string, enum: [watching, offer_made, owned], default: watching, description: CLOSED }
    PortfolioEntryUpdate:
      type: object
      properties:
        purchase_price: { $ref: '#/components/schemas/Money' }
        monthly_rental_income: { $ref: '#/components/schemas/Money' }
        monthly_expenses: { $ref: '#/components/schemas/Money' }
        outstanding_debt: { $ref: '#/components/schemas/Money' }
        monthly_mortgage_payment: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [watching, offer_made, owned], description: CLOSED }
    PortfolioEntry:
      allOf:
        - $ref: '#/components/schemas/PortfolioEntryInput'
        - type: object
          required: [status, added_at]
          properties:
            status: { type: string, enum: [watching, offer_made, owned], description: CLOSED }
            added_at: { type: string, format: date-time }
            current_value_estimate:
              allOf: [{ $ref: '#/components/schemas/Money' }]
              description: 'Server-computed from comparable listings. Absent (not null-as-zero) until enough comps exist to estimate. Never client-supplied.'
            current_value_estimate_computed_at:
              type: [string, 'null']
              format: date-time
              description: 'When the valuation job last ran for this entry — stamped even when the number did not change. Null only if valuation has never run. Use it to caveat ("as of …") or suppress projections built on a stale estimate.'
...
    SlotHold:
      type: object
      required: [id, listing_id, starts_at, ends_at, expires_at]
      properties:
        id: { type: string, format: uuid }
        listing_id: { type: string, format: uuid }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, description: Server-authoritative TTL }
    Appointment:
      type: object
      required: [id, state, starts_at, ends_at, kind]
      properties:
        id: { type: string, format: uuid }
        listing_id: { type: string, format: uuid }
        property_summary:
          type: object
          properties:
            address_line: { type: string, description: Masked until a purpose-bound grant applies }
            city: { type: string }
            location: { $ref: '#/components/schemas/GeoPoint' }
        state: { type: string, enum: [requested, hold_placed, dispatching, unstaffed, booked, confirmed, in_progress, completed, outcome_captured, cancelled, no_show, rescheduled], description: CLOSED }
        kind: { type: string, enum: [private, open_house] }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        agent: { $ref: '#/components/schemas/PartyRef' }
        viewer: { $ref: '#/components/schemas/PartyRef' }
        contact_reveal:
          type: [object, 'null']
          description: Present only inside an active purpose-bound access window.
          properties:
            window_ends_at: { type: string, format: date-time }
        cancellation_policy:
          type: object
          properties:
            free_until: { type: string, format: date-time }
            penalty_note_key: { type: string }
    PartyRef:
      type: object
      description: Masked party reference; PII revealed only under an active grant via the detail endpoints.
      properties:
        contact_id: { type: string, format: uuid }
        display_name: { type: string, description: May be masked (e.g. "S***") }
    CancellationResult:
      type: object
      required: [appointment]
      properties:
        appointment: { $ref: '#/components/schemas/Appointment' }
        penalty_applied: { type: boolean }
        penalty_note_key: { type: [string, 'null'] }
    AttendanceProofInput:
      type: object
      required: [method]
      properties:
        method: { type: string, enum: [geofence, one_time_code] }
        location: { $ref: '#/components/schemas/GeoPoint' }
        code: { type: string }
    AppointmentFeedbackInput:
      type: object
      properties:
        condition_rating: { type: integer, minimum: 1, maximum: 5 }
        price_opinion: { type: string, enum: [under, fair, over] }
        interest_level: { type: string, enum: [none, low, medium, high] }
        comments: { type: string, maxLength: 2000 }
        share_with_owner: { type: boolean, default: false }
...